requirements.txt
anthropic>=0.83.0
python-dotenv>=1.2.1
requests>=2.32.5
pandas>=3.0.1
beautifulsoup4>=4.14.3
yfinance>=1.2.0
jinja2>=3.1.6
feedparser>=6.0.12
lxml>=6.0.2
